# backend/docs/swagger.yaml
definitions:
...
  domain.Session:
    properties:
      created_at:
        type: string
      expires_at:
        type: string
      session_token:
        type: string
      status:
        type: string
      user_id:
        type: integer
    type: object
...
  usersHandler.ChangePasswordRequest:
    properties:
      current_password:
        minLength: 8
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
info:
  contact: {}
paths:
...
  /users/{user_id}:
    delete:
      consumes:
      - application/json
      description: Deletes an active session of a user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Session deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request URL, no user_id or session_id found.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Session not found
          schema:
            type: string
        "500":
          description: Failed to commit the transaction.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a user's active session
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Retrieves the user for a specific user ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Session status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of active sessions
          schema:
            items:
              $ref: '#/definitions/domain.Session'
            type: array
        "400":
          description: Invalid request URL, no user_id found.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Could not create users struct.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - Users
  /users/{user_id}/sessions:
    get:
      consumes:
      - application/json
      description: Retrieves the user for a specific user ID.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Session status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with a list of active sessions
          schema:
            items:
              $ref: '#/definitions/domain.Session'
            type: array
        "400":
          description: Invalid request URL, no user_id found.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Could not create users struct.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a user by ID
      tags:
      - Users
  /users/{user_id}/sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Deletes an active session of a user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Session deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request URL, no user_id or session_id found.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Session not found
          schema:
            type: string
        "500":
          description: Failed to commit the transaction.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a user's active session
      tags:
      - Users
  /users/password:
    patch:
      consumes:
      - application/json
      description: Updates the password of a user.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: New password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/usersHandler.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            type: string
        "400":
          description: New password cannot be the same as the current password.
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to commit the transaction.
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a user's password
      tags:
      - Users
swagger: "2.0"