# backend/internal/handler/usersHandler/changePasswordRequest.go
package usersHandler

type ChangePasswordRequest struct {
	CurrentPassword string `json:"current_password" validate:"required,min=8"`
	NewPassword     string `json:"new_password" validate:"required,min=8"`
}